+++ jest-puppeteer.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  server: {
-    command: 'npm run test:serve',
-    port: 1234,
-    launchTimeout: 50000
-  }
-}