+++ src/shims/perf_hooks.js
@@ -1,0 +1,3 @@
+module.exports = {
+  performance: typeof performance !== 'undefined' ? performance : Date
+}